Conflicts:
	features/events/syslog/src/main/java/org/opennms/netmgt/syslogd/BroadcastEventProcessor.java
	features/themes/dashboard-theme/src/main/java/VAADIN/themes/dashboard/styles.css
	opennms-doc/guide-admin/src/asciidoc/index.adoc

+++ features/vaadin-surveillance-views/src/main/java/org/opennms/features/vaadin/surveillanceviews/service/DefaultSurveillanceViewService.java
@@ -520,8 +520,8 @@ public class DefaultSurveillanceViewService implements SurveillanceViewService {
      * @return the computed availability
      */
 
-    private static Double calculateAvailability(int serviceCount, long downMillisCount) {
-        long upMillis = ((long) serviceCount * (24L * 60L * 60L * 1000L)) - downMillisCount;
+    private static Double calculateAvailability(long serviceCount, long downMillisCount) {
+        long upMillis = (serviceCount * (24L * 60L * 60L * 1000L)) - downMillisCount;
 
         return ((double) upMillis / (double) (serviceCount * (24 * 60 * 60 * 1000)));
     }
